Moved /net, /audio and /sound folders to /platform/
