+++ src/framework/components/audio-source/component.js
@@ -1,6 +1,6 @@
 import { Asset } from '../../../asset/asset.js';
 
-import { Channel3d } from '../../../audio/channel3d.js';
+import { Channel3d } from '../../../platform/audio/channel3d.js';
 
 import { Component } from '../component.js';
