+++ src/framework/components/audio-source/system.js
@@ -1,4 +1,4 @@
-import { Channel3d } from '../../../audio/channel3d.js';
+import { Channel3d } from '../../../platform/audio/channel3d.js';
 import { Debug } from '../../../core/debug.js';
 
 import { Entity } from '../../entity.js';
@@ -10,7 +10,7 @@ import { AudioSourceComponent } from './component.js';
 import { AudioSourceComponentData } from './data.js';
 
 /** @typedef {import('../../app-base.js').AppBase} AppBase */
-/** @typedef {import('../../../sound/manager.js').SoundManager} SoundManager */
+/** @typedef {import('../../../platform/sound/manager.js').SoundManager} SoundManager */
 
 const _schema = [
     'enabled',
