+++ src/framework/components/sound/slot.js
@@ -6,8 +6,8 @@ import { Vec3 } from '../../../core/math/vec3.js';
 
 import { Asset } from '../../../asset/asset.js';
 
-import { SoundInstance } from '../../../sound/instance.js';
-import { SoundInstance3d } from '../../../sound/instance3d.js';
+import { SoundInstance } from '../../../platform/sound/instance.js';
+import { SoundInstance3d } from '../../../platform/sound/instance3d.js';
 
 /** @typedef {import('./component.js').SoundComponent} SoundComponent */
