+++ src/platform/sound/instance.js
@@ -1,6 +1,6 @@
-import { EventHandler } from '../core/event-handler.js';
+import { EventHandler } from '../../core/event-handler.js';
 
-import { math } from '../core/math/math.js';
+import { math } from '../../core/math/math.js';
 
 import { hasAudioContext } from '../audio/capabilities.js';
