+++ src/platform/sound/listener.js
@@ -1,7 +1,7 @@
-import { Debug } from '../core/debug.js';
+import { Debug } from '../../core/debug.js';
 
-import { Vec3 } from '../core/math/vec3.js';
-import { Mat4 } from '../core/math/mat4.js';
+import { Vec3 } from '../../core/math/vec3.js';
+import { Mat4 } from '../../core/math/mat4.js';
 
 /** @typedef {import('./manager.js').SoundManager} SoundManager */
