+++ src/platform/sound/sound.js
