+++ src/resources/parser/texture/img.js
@@ -1,5 +1,5 @@
 import { path } from '../../../core/path.js';
-import { http } from '../../../net/http.js';
+import { http } from '../../../platform/net/http.js';
 
 import {
     PIXELFORMAT_R8_G8_B8, PIXELFORMAT_R8_G8_B8_A8, TEXHINT_ASSET
