+++ src/resources/sprite.js
@@ -1,6 +1,6 @@
 import { path } from '../core/path.js';
 
-import { http } from '../net/http.js';
+import { http } from '../platform/net/http.js';
 
 import { Sprite } from '../scene/sprite.js';
