+++ src/resources/texture-atlas.js
@@ -1,6 +1,6 @@
 import { path } from '../core/path.js';
 
-import { http } from '../net/http.js';
+import { http } from '../platform/net/http.js';
 
 import { Vec2 } from '../core/math/vec2.js';
 import { Vec4 } from '../core/math/vec4.js';
